Fix the error code to properly display
In a previous commit [1], an attempt was made to add the error code
to the resulting message. However, this did not work.

The correct way to get the error code is to override the result
of the error or warning variable.

This patch will now show the message as:
 21:1 warning bandit: B303 Use of insecure MD2, MD4, or MD5 hash function.

instead of:
 21:1 warning bandit: Medium Use of insecure MD2, MD4, or MD5 hash function.

[1] https://github.com/SublimeLinter/SublimeLinter-bandit/pull/3

Signed-off-by: Eric Brown <browne@vmware.com>